move study prefix to study file
