+++ src/services/dynamic-margin-calculation.ts
@@ -7,7 +7,7 @@
 import type { UUID } from 'node:crypto';
 import { DynamicMarginCalculationParametersInfos } from '../utils/types/dynamic-margin-calculation.type';
 import { backendFetch, backendFetchJson } from './utils';
-import { PREFIX_STUDY_SERVER_QUERIES } from './loadflow';
+import { PREFIX_STUDY_SERVER_QUERIES } from './study';
 
 function getDynamicMarginCalculationUrl() {
     return `${PREFIX_STUDY_SERVER_QUERIES}/v1/dynamic-margin-calculation/`;
