+++ src/services/security-analysis.ts
@@ -7,12 +7,12 @@
 
 import type { UUID } from 'node:crypto';
 import { backendFetch, backendFetchJson, safeEncodeURIComponent } from './utils';
-import { PREFIX_STUDY_QUERIES, PREFIX_STUDY_SERVER_QUERIES } from './loadflow';
+import { PREFIX_STUDY_SERVER_QUERIES } from './study';
 import { IdName, mapSecurityAnalysisParameters, SAParameters, SAParametersEnriched } from '../utils';
 import { fetchElementNames } from './directory';
 
 export const getStudyUrl = (studyUuid: UUID | null) =>
-    `${PREFIX_STUDY_QUERIES}/v1/studies/${safeEncodeURIComponent(studyUuid)}`;
+    `${PREFIX_STUDY_SERVER_QUERIES}/v1/studies/${safeEncodeURIComponent(studyUuid)}`;
 
 function getSecurityAnalysisUrl() {
     return `${PREFIX_STUDY_SERVER_QUERIES}/v1/security-analysis/`;
